pyproject.toml
[project]
name = "voice-agent"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.14"
dependencies = [
    "httpx>=0.28.1",
    "numpy>=1.26",
    "openai>=2.0.0,<3",
    "openai-agents[voice]>=0.13.6",
    "python-dotenv>=1.2.2",
    "rich>=15.0.0",
    "sounddevice>=0.5",
    "webrtcvad-wheels>=2.0.14",
]

[project.optional-dependencies]
local = [
    "mlx-audio[server]",
    "mlx-vlm",
]

[project.scripts]
voice-agent = "voice_agent.__main__:main"

[dependency-groups]
dev = [
    "pyright>=1.1.408",
    "ruff>=0.15.10",
]
